sandbox/oauth_reference/oauth_module/requirements-production.txt
# Production requirements for OAuth Streamlit app
# Includes Google Cloud dependencies for Secret Manager

# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0

# Data processing (for sample dashboard)
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud dependencies
google-cloud-secret-manager>=2.16.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Security and performance
cryptography>=41.0.0
gunicorn>=21.2.0

# Health checks and monitoring
psutil>=5.9.0